sh and .make

+++ scripts/deploy_dev_build.sh
@@ -5,6 +5,25 @@ die () {
   exit 1
 }
 
+# run commands will look repetitive; the only reason behind this is pure laziness and
+# just in case at some point I need to do something special with one of them
+
+run_compass() {
+    if hash compass 2>/dev/null; then
+        compass "$@"
+    else
+        die "Unable to find compass. Make sure is instlled."
+    fi
+}
+
+run_drush() {
+    if hash drush 2>/dev/null; then
+        drush "$@"
+    else
+        die "Unable to find drush. Make sure is instlled."
+    fi
+}
+
 [ "$#" -eq 1 ] || die "Please provide make file"
 
 SOURCE="${BASH_SOURCE[0]}"
@@ -29,11 +48,13 @@ LIBRARY="$(dirname "$DIR")"/lib
 [ -d $BUILD_DIR ] || die "Build directory $BUILD_DIR does not exist" 
 
 echo Preparing new site using $1
-drush make --prepare-install $1 $BUILD_DIR/$BUILD_NAME
-  
+run_drush make --prepare-install $1 $BUILD_DIR/$BUILD_NAME
+
+# i need to add a test to check if we have all we need
+
 echo Install new site
 cd $BUILD_DIR/$BUILD_NAME
-drush site-install $DRUPAL_INSTALL_PROFILE_NAME --site-name=$DRUPAL_SITE_NAME --account-name=$DRUPAL_ACCOUNT_NAME --account-mail=$DRUPAL_ACCOUNT_MAIL --site-mail=$DRUPAL_SITE_MAIL --db-url=$DRUPAL_SITE_DB_TYPE://$DRUPAL_SITE_DB_USER:$DRUPAL_SITE_DB_PASS@$DRUPAL_SITE_DB_ADDRESS/$DRUPAL_DB_NAME
+run_drush site-install $DRUPAL_INSTALL_PROFILE_NAME --site-name=$DRUPAL_SITE_NAME --account-name=$DRUPAL_ACCOUNT_NAME --account-mail=$DRUPAL_ACCOUNT_MAIL --site-mail=$DRUPAL_SITE_MAIL --db-url=$DRUPAL_SITE_DB_TYPE://$DRUPAL_SITE_DB_USER:$DRUPAL_SITE_DB_PASS@$DRUPAL_SITE_DB_ADDRESS/$DRUPAL_DB_NAME
   
 echo Linking build name $BUILD_NAME
 
@@ -43,11 +64,14 @@ site_dirs=(modules themes)
 for site_dir in "${site_dirs[@]}"
   do
     for dir in $LIBRARY/${site_dir}/*
-      do 
+      do
         base=${dir##*/}
-        rm -rf $BUILD_DIR/$BUILD_NAME/sites/all/${site_dir}/${base}
-        ln -s $LIBRARY/${site_dir}/${base} $BUILD_DIR/$BUILD_NAME/sites/all/${site_dir}/${base}
-    done
+        if [ -d $BUILD_DIR/$BUILD_NAME/sites/all/${site_dir}/${base} ] && [ -d $LIBRARY/${site_dir}/${base} ]
+          then
+            rm -rf $BUILD_DIR/$BUILD_NAME/sites/all/${site_dir}/${base}
+            ln -s $LIBRARY/${site_dir}/${base} $BUILD_DIR/$BUILD_NAME/sites/all/${site_dir}/${base}
+        fi  
+  done
 done
 
 # find profile and symlink them to the repo code
@@ -62,20 +86,33 @@ done
 imagemagick_convert_path=`which convert`
 
 # Set $imagemagick_convert_path as imagemagick convert path
-drush -r $BUILD_DIR/$BUILD_NAME vset imagemagick_convert $imagemagick_convert_path
+run_drush -r $BUILD_DIR/$BUILD_NAME vset imagemagick_convert $imagemagick_convert_path
 
-echo Building OpenLayers
+if [ -d $BUILD_DIR/$BUILD_NAME/sites/all/modules/dlts_image ]
+  then 
 
-cp $BUILD_DIR/$BUILD_NAME/sites/all/modules/dlts_image/js/openlayers/books.cfg $BUILD_DIR/$BUILD_NAME/sites/all/libraries/openlayers/build/books.cfg
-cp $BUILD_DIR/$BUILD_NAME/sites/all/modules/dlts_image/js/openlayers/DLTS.js $BUILD_DIR/$BUILD_NAME/sites/all/libraries/openlayers/lib/OpenLayers/DLTS.js
-cp $BUILD_DIR/$BUILD_NAME/sites/all/modules/dlts_image/js/openlayers/DLTSZoomIn.js $BUILD_DIR/$BUILD_NAME/sites/all/libraries/openlayers/lib/OpenLayers/Control/DLTSZoomIn.js
-cp $BUILD_DIR/$BUILD_NAME/sites/all/modules/dlts_image/js/openlayers/DLTSZoomOut.js $BUILD_DIR/$BUILD_NAME/sites/all/libraries/openlayers/lib/OpenLayers/Control/DLTSZoomOut.js
-cp $BUILD_DIR/$BUILD_NAME/sites/all/modules/dlts_image/js/openlayers/DLTSZoomPanel.js $BUILD_DIR/$BUILD_NAME/sites/all/libraries/openlayers/lib/OpenLayers/Control/DLTSZoomPanel.js
-cp $BUILD_DIR/$BUILD_NAME/sites/all/modules/dlts_image/js/openlayers/DLTSZoomOutPanel.js $BUILD_DIR/$BUILD_NAME/sites/all/libraries/openlayers/lib/OpenLayers/Control/DLTSZoomOutPanel.js
-cp $BUILD_DIR/$BUILD_NAME/sites/all/modules/dlts_image/js/openlayers/DLTSZoomInPanel.js $BUILD_DIR/$BUILD_NAME/sites/all/libraries/openlayers/lib/OpenLayers/Control/DLTSZoomInPanel.js
-cp $BUILD_DIR/$BUILD_NAME/sites/all/modules/dlts_image/js/openlayers/DLTSScrollWheel.js $BUILD_DIR/$BUILD_NAME/sites/all/libraries/openlayers/lib/OpenLayers/Control/DLTSScrollWheel.js
-cp $BUILD_DIR/$BUILD_NAME/sites/all/modules/dlts_image/js/openlayers/DLTSMouseWheel.js $BUILD_DIR/$BUILD_NAME/sites/all/libraries/openlayers/lib/OpenLayers/Handler/DLTSMouseWheel.js
-cp $BUILD_DIR/$BUILD_NAME/sites/all/modules/dlts_image/js/openlayers/OpenURL.js $BUILD_DIR/$BUILD_NAME/sites/all/libraries/openlayers/lib/OpenLayers/Layer/OpenURL.js
+    echo Building OpenLayers
+
+    cp $BUILD_DIR/$BUILD_NAME/sites/all/modules/dlts_image/js/openlayers/books.cfg $BUILD_DIR/$BUILD_NAME/sites/all/libraries/openlayers/build/books.cfg
+    cp $BUILD_DIR/$BUILD_NAME/sites/all/modules/dlts_image/js/openlayers/DLTS.js $BUILD_DIR/$BUILD_NAME/sites/all/libraries/openlayers/lib/OpenLayers/DLTS.js
+
+    cp $BUILD_DIR/$BUILD_NAME/sites/all/modules/dlts_image/js/openlayers/DLTSZoomIn.js $BUILD_DIR/$BUILD_NAME/sites/all/libraries/openlayers/lib/OpenLayers/Control/DLTSZoomIn.js
+
+    cp $BUILD_DIR/$BUILD_NAME/sites/all/modules/dlts_image/js/openlayers/DLTSZoomOut.js $BUILD_DIR/$BUILD_NAME/sites/all/libraries/openlayers/lib/OpenLayers/Control/DLTSZoomOut.js
+
+    cp $BUILD_DIR/$BUILD_NAME/sites/all/modules/dlts_image/js/openlayers/DLTSZoomPanel.js $BUILD_DIR/$BUILD_NAME/sites/all/libraries/openlayers/lib/OpenLayers/Control/DLTSZoomPanel.js
+
+    cp $BUILD_DIR/$BUILD_NAME/sites/all/modules/dlts_image/js/openlayers/DLTSZoomOutPanel.js $BUILD_DIR/$BUILD_NAME/sites/all/libraries/openlayers/lib/OpenLayers/Control/DLTSZoomOutPanel.js
+
+    cp $BUILD_DIR/$BUILD_NAME/sites/all/modules/dlts_image/js/openlayers/DLTSZoomInPanel.js $BUILD_DIR/$BUILD_NAME/sites/all/libraries/openlayers/lib/OpenLayers/Control/DLTSZoomInPanel.js
+
+    cp $BUILD_DIR/$BUILD_NAME/sites/all/modules/dlts_image/js/openlayers/DLTSScrollWheel.js $BUILD_DIR/$BUILD_NAME/sites/all/libraries/openlayers/lib/OpenLayers/Control/DLTSScrollWheel.js
+
+    cp $BUILD_DIR/$BUILD_NAME/sites/all/modules/dlts_image/js/openlayers/DLTSMouseWheel.js $BUILD_DIR/$BUILD_NAME/sites/all/libraries/openlayers/lib/OpenLayers/Handler/DLTSMouseWheel.js
+
+    cp $BUILD_DIR/$BUILD_NAME/sites/all/modules/dlts_image/js/openlayers/OpenURL.js $BUILD_DIR/$BUILD_NAME/sites/all/libraries/openlayers/lib/OpenLayers/Layer/OpenURL.js
+
+fi
 
 cd $BUILD_DIR/$BUILD_NAME/sites/all/libraries/openlayers/build
 
@@ -84,14 +121,14 @@ cd $BUILD_DIR/$BUILD_NAME/sites/all/libraries/openlayers/build
 cd $DIR
 
 if [ -f $BUILD_DIR/$BUILD_NAME/sites/all/libraries/openlayers/build/OpenLayers.js ]
-then
-  drush -r $BUILD_DIR/$BUILD_NAME vset dlts_image_openlayers_source sites/all/libraries/openlayers/build/OpenLayers.js
+  then
+    run_drush -r $BUILD_DIR/$BUILD_NAME vset dlts_image_openlayers_source sites/all/libraries/openlayers/build/OpenLayers.js
 fi
 
 echo Generating CSS files using compass
-compass compile --force $LIBRARY/themes/dlts_book
+run_compass compile --force $LIBRARY/themes/dlts_book
 
-drush -r $BUILD_DIR/$BUILD_NAME scr $DIR/dummy_book/import_dummy_book.php
+run_drush -r $BUILD_DIR/$BUILD_NAME scr $DIR/dummy_book/import_dummy_book.php
 
 cd $DIR
